[RAM-4] Refactor GameDataConverter subclasses into separate files
