+++ Assets/Source/Serialization/Data/BattlefieldItemData.cs
@@ -1,6 +1,6 @@
 using System;
 using Source.Logic;
-using Source.Logic.Data;
+using Source.Logic.State;
 
 namespace Source.Serialization.Data
 {
