+++ Assets/Source/Serialization/GameStateLoader.cs
@@ -10,6 +10,7 @@ namespace Source.Serialization
     {
         [Header("Dependencies")]
         [SerializeField] private TextAsset gameStateJsonAsset;
+        [SerializeField] private GameResources gameResources;
 
         public GameState GameState => gameState;
 
@@ -26,7 +27,7 @@ namespace Source.Serialization
         { 
             gameData = jsonDataService.LoadData<GameData>(Encoding.UTF8.GetBytes(gameStateJsonAsset.text), false);
             var converter = new GameDataConverter();
-            gameState = converter.Convert(gameData);
+            gameState = converter.Convert(gameData, gameResources);
             
             Debug.Log($"Game state: {gameData}");
         }
@@ -35,7 +36,7 @@ namespace Source.Serialization
         {
             gameData = jsonDataService.LoadData<GameData>(relativePath, false);
             var converter = new GameDataConverter();
-            gameState = converter.Convert(gameData);
+            gameState = converter.Convert(gameData, gameResources);
         }
 
         public void Save(GameState gameState)
